advancing cpp09, parsing csv ok

+++ cpp09/ex00/BitcoinExchange.hpp
@@ -5,12 +5,14 @@
 #include <map>
 #include <fstream>
 #include <string>
+#include <exception>
+#include <stdlib.h>
 
 class BitcoinExchange 
 {
 	private:
 	public:
-		std::map<std::string, float> _values;
+		std::map<std::string, double> _values;
 		BitcoinExchange();
 		BitcoinExchange(const BitcoinExchange &ref);
 		BitcoinExchange &operator=(const BitcoinExchange &ref);
